4090_LLM/NoAug_Criteria_Evidence/pyproject.toml
[project]
name = "noaug-criteria-evidence"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "torch>=2.2",
  "transformers>=4.44",
  "tokenizers>=0.15",
  "optuna~=4.5.0",
  "mlflow>=2.14",
  "typer>=0.12",
  "pydantic>=2.8",
  "scikit-learn>=1.4",
  "datasets>=2.20",
  "evaluate>=0.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=4.1",
  "ruff>=0.6",
  "black>=24.0",
  "mypy>=1.9",
]

[project.scripts]
psy-agents = "psy_agents_noaug.cli:main"

[tool.poetry]
name = "psy_agents_noaug"
version = "0.1.0"
description = "NO-AUG baseline for Criteria and Evidence extraction from clinical text"
authors = ["Research Team"]
readme = "README.md"
packages = [{include = "psy_agents_noaug", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = ">=2.2"
transformers = ">=4.44"
hydra-core = "^1.3.0"
mlflow = ">=2.14"
optuna = "~=4.5.0"
pandas = "^2.0.0"
numpy = ">=1.26"
scikit-learn = ">=1.4"
datasets = ">=2.20"
tokenizers = ">=0.15"
typer = ">=0.12"
pydantic = ">=2.8"
evaluate = ">=0.4"
nlpaug = "^1.1.11"  # For comparison only - NOT used in this repo
peft = "^0.10.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
ruff = ">=0.6"
black = ">=24.0"
isort = "^5.12.0"
pytest = ">=8.0"
pytest-cov = ">=4.1.0"
pytest-mock = "^3.11.1"
pre-commit = "^3.3.0"
bandit = "^1.7.5"
mypy = ">=1.9"

[tool.poetry.scripts]
psy-agents = "psy_agents_noaug.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by black)
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]

[tool.coverage.run]
source = ["src/psy_agents_noaug"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/conftest.py",
]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
precision = 2
show_missing = true

[tool.coverage.html]
directory = "htmlcov"

[tool.bandit]
exclude_dirs = ["tests", ".venv", "venv"]
skips = ["B101"]  # assert_used - we use asserts in tests
